made it more responsive

+++ src/pages/Checkout.js
@@ -48,14 +48,15 @@ function Checkout() {
   }
 
   async function handleRemove(e, item) {
-    await dispatch(deleteItemFromCartAsync(item.id))
+    if (window.confirm('Are you sure you want to remove an item from the cart ?'))
+      await dispatch(deleteItemFromCartAsync(item.id))
   }
 
   async function handleOrder() {
     if (items && items.length > 0 && paymentMethod && selectedAddress) {
       const order = { items, totalAmount, totalItems, loggedInUserId: user.id, paymentMethod, selectedAddress, status: 'pending' };
       if (order.paymentMethod === 'card') {
-        navigate('/stripe-checkout', {state:order});
+        navigate('/stripe-checkout', { state: order });
       }
       else if (order.paymentMethod === 'cash') {
         const response = await dispatch(addOrderAsync(order));
@@ -80,11 +81,7 @@ function Checkout() {
 
   return (
     <Navbar>
-      {/* {currentOrder !== undefined && currentOrder.paymentMethod === 'cash' && <Navigate to={`/orderSuccess/${currOrder.id}`} replace={true} />} */}
-      {/* TODO: BHAI Y CHEEZ GALAT HAI YAHAN PE VO WALA LOGIC LGEGA KI JAB PAYMENT SUCCESSFULL HO JAYE TB USER KA ORDER CREATE HO VRNA NA HO */}
-      {/* {currentOrder !== undefined && currentOrder.paymentMethod === 'card' && <Navigate to={`/stripe-checkout`} replace={true} />} */}
-
-      <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:flex p-3 space-x-5 font-serif lg:px-8">
+      <div className="mx-auto max-w-7xl sm:px-6 lg:flex space-x-5 font-serif lg:px-8">
         <div className='lg:w-1/2 text-start bg-white p-5'>
           <div className="space-y-12">
 
@@ -365,106 +362,114 @@ function Checkout() {
 
 
         </div>
-        <div className='lg:w-1/2'> <div className="mx-auto mt-10 z-10 overflow-x-hidden bg-white max-w-7xl py-6 sm:px-6 lg:px-8"><div className="mt-8">
-          <h2 className="text-4xl font-serif font-bold text-blue-700 underline ">My cart</h2>
-          <div className="flow-root">
-            <ul role="list" className="-my-6 p-5 divide-y divide-gray-200">
-              {items && items.map((item) => (
-                <li key={item.id} className="flex py-6">
-                  <div className="h-24 w-24 flex-shrink-0 overflow-hidden rounded-md border border-gray-200">
-                    <img
-                      src={item.productId.thumbnail}
-                      alt={item.productId.title}
-                      className="h-full w-full object-cover object-center"
-                    />
-                  </div>
 
-                  <div className="ml-4 flex flex-1 flex-col">
-                    <div>
-                      <div className="flex justify-between text-base font-medium text-gray-900">
-                        <h3>
-                          <a href={item.productId.title}>{item.productId.title}</a>
-                        </h3>
-                        <p className="ml-4">${discountedPrice(item.productId) * item.quantity}</p>
+        <div className='lg:w-1/2'>
+          <div className="mt-10 z-10 overflow-x-hidden bg-white max-w-7xl  sm:px-6 lg:px-8">
+            <div className="mt-8">
+              <h2 className="text-4xl font-serif font-bold text-blue-700 underline  text-center">My cart</h2>
+              <div className="flow-root ">
+                <ul role="list" className="-my-6 p-5 divide-y divide-gray-200">
+                  {items && items.map((item) => (
+                    <li key={item.id} className="sm:flex text-center py-6">
+                      {/* IMAGE */}
+                      <div className="h-24 w-full sm:w-24 flex-shrink-0 overflow-hidden rounded-md ">
+                        <img
+                          src={item.productId.thumbnail}
+                          alt={item.productId.title}
+                          className="h-full m-auto "
+                        />
                       </div>
-                      <p className="text-end line-through">${item.productId.price * item.quantity}</p>
 
-                      <p className="mt-1 text-sm text-gray-500">Stock Available: {item.productId.stock - item.quantity}</p>
-                      <p className="mt-1 text-sm text-gray-500">Quantity Selected: {item.quantity}</p>
-                    </div>
-                    <div className="flex flex-1 mt-2 items-end justify-between text-sm">
-                      <div className=" text-black font-bold ">Qty
-                        <div className="inline ml-2">
-
-                          <input
-                            type="number"
-                            value={item.quantity}
-                            onChange={(e) => handleQuantity(e, item)}
-                            min={1}
-                            max={item.productId.stock}  // Set the maximum value to the available stock
-                          />
+                      {/* BAAKI DETAILS WALA TAG */}
+                      <div className="sm:p-3 flex  flex-1 flex-col">
+                        <div>
+                          <div className="sm:flex justify-between text-base font-medium text-gray-900">
+                            <h3>
+                              <a href={item.productId.title}> { } {item.productId.title}</a>
+                            </h3>
+                            <p >${discountedPrice(item.productId) * item.quantity}</p>
+                          </div>
+                          <p className=" sm:text-end font-semibold line-through">${item.productId.price * item.quantity}</p>
+                          <p className="mt-1 text-sm sm:text-start text-gray-500">Stock Available: {item.productId.stock - item.quantity}</p>
+                          <p className="mt-1 text-sm sm:text-start text-gray-500">Quantity Selected: {item.quantity}</p>
                         </div>
-                      </div>
+
+                        <div className="sm:flex flex-1 items-end mt-2 justify-between text-sm">
+                          <div className=" text-black font-bold ">Qty
+                            <div className="inline mt-2 ml-2">
+                              <input
+                                type="number"
+                                onChange={(e) => handleQuantity(e, item)}
+                                value={item.quantity}
+                                min={1}
+                                max={item.productId.stock}  // Set the maximum value to the available stock
+                              />
+                            </div>
+                          </div>
 
 
-                      <div className="flex">
-                        <button
-                          onClick={(e) => handleRemove(e, item)}
-                          type="button"
-                          className="font-medium text-indigo-600 hover:text-indigo-500"
-                        >
-                          Remove
-                        </button>
+
+                          <div className="">
+                            <button
+                              onClick={(e) => handleRemove(e, item)}
+                              type="button"
+                              className="font-medium  text-indigo-600 mt-2 hover:text-indigo-500"
+                            >
+                              Remove Item
+                            </button>
+
+                          </div>
+                        </div>
+
                       </div>
-                    </div>
-                  </div>
-                </li>
-              ))}
-            </ul>
+                    </li>
+                  ))}
+                </ul>
+              </div>
+            </div>
+
+            {totalItems > 0 ? (<div className="border-t border-gray-200 px-4 py-6 sm:px-6">
+              <div className="flex justify-between text-base font-medium text-gray-900">
+                <p>Subtotal</p>
+                <p>
+                  ${totalAmount}
+                </p>
+              </div>
+              <div className="flex justify-between text-base font-medium text-gray-900">
+                <p>Total Items</p>
+                <p>
+                  {totalItems} items
+                </p>
+              </div>
+                <div className="mt-6">
+                  <button
+                  onClick={handleOrder}
+                    className="flex text-center m-auto items-center justify-center rounded-md border border-transparent bg-indigo-600 px-6 py-3 text-base font-medium text-white shadow-sm hover:bg-indigo-700"
+                  >
+                    Place Order
+                  </button>
+                </div>
+              <div className="mt-6 flex justify-center text-center text-sm text-gray-500">
+                <p>
+                  or{' '}
+                  <Link to='/'>
+                    <button
+                      type="button"
+                      className="font-medium text-indigo-600 hover:text-indigo-500"
+                    // onClick={() => setOpen(false)}
+                    >
+                      Continue Shopping
+                      <span aria-hidden="true"> &rarr;</span>
+                    </button>
+                  </Link>
+
+                </p>
+              </div>
+            </div>) : <div className="flex flex-col justify-start items-start mt-2">OOPS ! Card is empty
+              <button className="font-serif mt-2 bg-gray-200 px-2 py-1 hover:text-yellow-900 rounded-sm" onClick={() => navigate('/login')}>Start Adding items</button>
+            </div>}
           </div>
         </div>
-
-          {totalItems > 0 ? (<div className="border-t border-gray-200 px-4 py-6 sm:px-6">
-            <div className="flex justify-between text-base font-medium text-gray-900">
-              <p>Subtotal</p>
-              <p>
-                ${totalAmount}
-              </p>
-            </div>
-            <div className="flex justify-between text-base font-medium text-gray-900">
-              <p>Total Items</p>
-              <p>
-                {totalItems} items
-              </p>
-            </div><p className="mt-0.5 text-sm text-start text-black">Shipping and taxes calculated at checkout.</p>
-            <div className="mt-6">
-              <button
-                onClick={handleOrder}
-                className="flex text-center m-auto items-center justify-center rounded-md border border-transparent bg-indigo-600 px-6 py-3 text-base font-medium text-white shadow-sm hover:bg-indigo-700"
-              >
-                Place Order
-              </button>
-            </div>
-            <div className="mt-6 flex justify-center text-center text-sm text-gray-500">
-              <p>
-                or{' '}
-                <Link to='/'>
-                  <button
-                    type="button"
-                    className="font-medium text-indigo-600 hover:text-indigo-500"
-                  // onClick={() => setOpen(false)}
-                  >
-                    Continue Shopping
-                    <span aria-hidden="true"> &rarr;</span>
-                  </button>
-                </Link>
-
-              </p>
-            </div>
-          </div>) : <div className="flex flex-col justify-start items-start mt-2">OOPS ! Card is empty
-            <button className="font-serif mt-2 bg-gray-200 px-2 py-1 hover:text-yellow-900 rounded-sm" onClick={() => navigate('/login')}>Start Adding items</button>
-          </div>}
-        </div></div>
       </div>
     </Navbar>
   )
